Restructure plugins
