+++ packages/plugins/tests/web/url.test.ts
@@ -14,97 +14,97 @@ import {
   toJSON,
   toJSONAsync,
 } from 'seroval';
-import URLSearchParamsPlugin from '.';
+import URLPlugin from '../../web/url';
 
-const EXAMPLE = new URLSearchParams('hello=world&foo=bar');
+const EXAMPLE = new URL('https://github.com/lxsmnsyc/seroval?hello=world');
 
-describe('URLSearchParams', () => {
+describe('URL', () => {
   describe('serialize', () => {
-    it('supports URLSearchParams', () => {
+    it('supports URL', () => {
       const result = serialize(EXAMPLE, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(result).toMatchSnapshot();
       const back = deserialize<typeof EXAMPLE>(result);
-      expect(back).toBeInstanceOf(URLSearchParams);
+      expect(back).toBeInstanceOf(URL);
       expect(String(back)).toBe(String(EXAMPLE));
     });
   });
   describe('serializeAsync', () => {
-    it('supports URLSearchParams', async () => {
+    it('supports URL', async () => {
       const result = await serializeAsync(Promise.resolve(EXAMPLE), {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(result).toMatchSnapshot();
       const back = await deserialize<Promise<typeof EXAMPLE>>(result);
-      expect(back).toBeInstanceOf(URLSearchParams);
+      expect(back).toBeInstanceOf(URL);
       expect(String(back)).toBe(String(EXAMPLE));
     });
   });
   describe('toJSON', () => {
-    it('supports URLSearchParams', () => {
+    it('supports URL', () => {
       const result = toJSON(EXAMPLE, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(JSON.stringify(result)).toMatchSnapshot();
       const back = fromJSON<typeof EXAMPLE>(result, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
-      expect(back).toBeInstanceOf(URLSearchParams);
+      expect(back).toBeInstanceOf(URL);
       expect(String(back)).toBe(String(EXAMPLE));
     });
   });
   describe('toJSONAsync', () => {
-    it('supports URLSearchParams', async () => {
+    it('supports URL', async () => {
       const result = await toJSONAsync(Promise.resolve(EXAMPLE), {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(JSON.stringify(result)).toMatchSnapshot();
-      const back = await fromJSON<Promise<typeof EXAMPLE>>(result, {
-        plugins: [URLSearchParamsPlugin],
+      const back = await fromJSON<Promise<URL>>(result, {
+        plugins: [URLPlugin],
       });
-      expect(back).toBeInstanceOf(URLSearchParams);
+      expect(back).toBeInstanceOf(URL);
       expect(String(back)).toBe(String(EXAMPLE));
     });
   });
   describe('crossSerialize', () => {
-    it('supports URLSearchParams', () => {
+    it('supports URL', () => {
       const result = crossSerialize(EXAMPLE, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(result).toMatchSnapshot();
     });
     describe('scoped', () => {
-      it('supports URLSearchParams', () => {
+      it('supports URL', () => {
         const result = crossSerialize(EXAMPLE, {
-          plugins: [URLSearchParamsPlugin],
           scopeId: 'example',
+          plugins: [URLPlugin],
         });
         expect(result).toMatchSnapshot();
       });
     });
   });
   describe('crossSerializeAsync', () => {
-    it('supports URLSearchParams', async () => {
+    it('supports URL', async () => {
       const result = await crossSerializeAsync(Promise.resolve(EXAMPLE), {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(result).toMatchSnapshot();
     });
     describe('scoped', () => {
-      it('supports URLSearchParams', async () => {
+      it('supports URL', async () => {
         const result = await crossSerializeAsync(Promise.resolve(EXAMPLE), {
-          plugins: [URLSearchParamsPlugin],
           scopeId: 'example',
+          plugins: [URLPlugin],
         });
         expect(result).toMatchSnapshot();
       });
     });
   });
   describe('crossSerializeStream', () => {
-    it('supports URLSearchParams', async () => new Promise<void>((resolve, reject) => {
+    it('supports URL', async () => new Promise<void>((resolve, reject) => {
       crossSerializeStream(Promise.resolve(EXAMPLE), {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
         onSerialize(data) {
           expect(data).toMatchSnapshot();
         },
@@ -117,9 +117,9 @@ describe('URLSearchParams', () => {
       });
     }));
     describe('scoped', () => {
-      it('supports URLSearchParams', async () => new Promise<void>((resolve, reject) => {
+      it('supports URL', async () => new Promise<void>((resolve, reject) => {
         crossSerializeStream(Promise.resolve(EXAMPLE), {
-          plugins: [URLSearchParamsPlugin],
+          plugins: [URLPlugin],
           scopeId: 'example',
           onSerialize(data) {
             expect(data).toMatchSnapshot();
@@ -135,37 +135,37 @@ describe('URLSearchParams', () => {
     });
   });
   describe('toJSON', () => {
-    it('supports URLSearchParams', () => {
+    it('supports URL', () => {
       const result = toCrossJSON(EXAMPLE, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(JSON.stringify(result)).toMatchSnapshot();
       const back = fromCrossJSON<typeof EXAMPLE>(result, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
         refs: new Map(),
       });
-      expect(back).toBeInstanceOf(URLSearchParams);
+      expect(back).toBeInstanceOf(URL);
       expect(String(back)).toBe(String(EXAMPLE));
     });
   });
   describe('toJSONAsync', () => {
-    it('supports URLSearchParams', async () => {
+    it('supports URL', async () => {
       const result = await toCrossJSONAsync(Promise.resolve(EXAMPLE), {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
       });
       expect(JSON.stringify(result)).toMatchSnapshot();
       const back = await fromCrossJSON<Promise<typeof EXAMPLE>>(result, {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
         refs: new Map(),
       });
-      expect(back).toBeInstanceOf(URLSearchParams);
+      expect(back).toBeInstanceOf(URL);
       expect(String(back)).toBe(String(EXAMPLE));
     });
   });
   describe('toCrossJSONStream', () => {
-    it('supports URLSearchParams', async () => new Promise<void>((resolve, reject) => {
+    it('supports URL', async () => new Promise<void>((resolve, reject) => {
       toCrossJSONStream(Promise.resolve(EXAMPLE), {
-        plugins: [URLSearchParamsPlugin],
+        plugins: [URLPlugin],
         onParse(data) {
           expect(JSON.stringify(data)).toMatchSnapshot();
         },
