+++ packages/plugins/tsconfig.eslint.json
@@ -1,6 +1,6 @@
 {
   "exclude": ["node_modules"],
-  "include": ["web"],
+  "include": ["web", "tests"],
   "compilerOptions": {
     "module": "ESNext",
     "lib": ["ESNext", "DOM"],
