+++ packages/plugins/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "exclude": ["node_modules"],
-  "include": ["web"],
+  "include": ["web", "tests/web/url.test.ts", "tests/web/url-search-params.test.ts"],
   "compilerOptions": {
     "module": "ESNext",
     "lib": ["ESNext", "DOM"],
